requirements.txt
streamlit
google-oauth2-tool